back/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "motor>=3.4",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "pillow>=10.0",
  "torch>=2.0",
  "torchvision>=0.15",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.5",
  "argon2-cffi>=21.3.0",
  "pydantic[email]>=2.6",
]

[tool.uvicorn]
factory = false
